% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer-package.R
\docType{package}
\name{cer-package}
\alias{cer}
\alias{cer-package}
\title{cer: Interactively Run CER Commands with Slurm}
\description{
Interactively Run CER (Climate Econometrics of Regional level) Commands with Slurm.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
