% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridFormat.R
\name{createGridFormat}
\alias{createGridFormat}
\title{GridFormat class}
\usage{
createGridFormat(degStep, lonFirst, lonIncreasing, latIncreasing)
}
\arguments{
\item{degStep}{The degree step of the grid.}

\item{lonFirst}{Whether the longitude is the first dimension.}

\item{lonIncreasing}{Whether the longitude values are in increasing order.}

\item{latIncreasing}{Whether the latitude values are in increasing order.}
}
\description{
A class to store the format of a longitude - latitude grid of the globe.
}
\examples{
createGridFormat(1, TRUE, TRUE, TRUE)
}
