% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EbmNetCdf-package.R
\docType{package}
\name{EbmNetCdf-package}
\alias{EbmNetCdf}
\alias{EbmNetCdf-package}
\title{EbmNetCdf: Deal with NetCDF files}
\description{
Deal with NetCDF files.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
