% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-package.R
\docType{package}
\name{run-package}
\alias{run}
\alias{run-package}
\title{run: Interactively Run Commands with Slurm}
\description{
Interactively Run Commands with Slurm.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
